+++ packages/preact-query-persist-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/preact-query-persist-client
 
+## 5.100.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/preact-query@5.100.11
+  - @tanstack/query-persist-client-core@5.100.11
+
 ## 5.100.10
 
 ### Patch Changes