+++ packages/solid-query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tanstack/solid-query
 
+## 5.100.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.100.11
+
 ## 5.100.10
 
 ### Patch Changes