+++ packages/svelte-query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tanstack/svelte-query
 
+## 6.1.30
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.100.11
+
 ## 6.1.29
 
 ### Patch Changes